feat: Add support for image extraction in LlamaParse
This commit adds a new boolean property `ExtractImages` to the `Configuration` class, allowing users to specify whether they want images to be extracted during parsing. In the `LlamaParse` class, if `ExtractImages` is set to true, the method `LoadImagesAsync()` is called to load and process images asynchronously. The extracted images are then added as separate documents with null content and copied metadata.

The intent of this change is to provide users with the ability to extract images along with text during parsing. This area has been tested by adding test cases that verify correct image extraction and document creation when `ExtractImages` is enabled.